Maps.kt menambahkan rute

+++ app/src/main/java/com/bornewtech/mitrapesaing/data/maps/LatLng.kt
@@ -1,7 +1,8 @@
 package com.bornewtech.mitrapesaing.data.maps
 
-data class LatLng(val cluster: Int = 0,
-                  val jmlh: Int = 0,
-                  val lat: Double = 0.0,
-                  val lng: Double = 0.0)
+data class LatLng(
+    val cluster: Double = 0.0,
+    val jmlh: Double = 0.0,
+    val lat: Double = 0.0,
+    val lng: Double = 0.0)
